scripts/requirements.txt
paramiko==3.4.0
pyyaml==6.0.1
python-dotenv==1.0.0